# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Asset Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PositionItem {
  /** Position/Number identifier (e.g., "01 / 1") */
  position: string;
  quantity?: string;
  /** Width in mm */
  width?: string;
  /** Height in mm */
  height?: string;
  /** Area in m²/Pcs. */
  area?: string;
  /** Perimeter in m/Pcs. */
  perimeter?: string;
  /** Price in SAR */
  price?: string;
  /** Total price in SAR */
  total?: string;
  /** Base64 data URL of the QR code image */
  qrDataUrl: string;
}

export interface ProcessResult {
  /** ID to use for downloading the processed file */
  fileId: string;
  positions: PositionItem[];
  projectName?: string;
  date?: string;
  totalPositions: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ProcessDocumentBody = {
  /** The Word (.docx) file to process */
  file: Blob;
  /** Optional base URL prefix for QR codes */
  qrBaseUrl?: string;
};