# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QR Asset Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: qr
    description: QR code generation from Orgadata documents
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /qr/process:
    post:
      operationId: processDocument
      tags: [qr]
      summary: Process Orgadata Word document
      description: Upload a Word document, parse positions, generate QR codes, return new document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The Word (.docx) file to process
                qrBaseUrl:
                  type: string
                  description: Optional base URL prefix for QR codes
              required:
                - file
      responses:
        "200":
          description: Processed document with QR codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /qr/download/{fileId}:
    get:
      operationId: downloadDocument
      tags: [qr]
      summary: Download processed document
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The processed Word document
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProcessResult:
      type: object
      properties:
        fileId:
          type: string
          description: ID to use for downloading the processed file
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionItem"
        projectName:
          type: string
        date:
          type: string
        totalPositions:
          type: integer
      required:
        - fileId
        - positions
        - totalPositions
    PositionItem:
      type: object
      properties:
        position:
          type: string
          description: Position/Number identifier (e.g., "01 / 1")
        quantity:
          type: string
        width:
          type: string
          description: Width in mm
        height:
          type: string
          description: Height in mm
        area:
          type: string
          description: Area in m²/Pcs.
        perimeter:
          type: string
          description: Perimeter in m/Pcs.
        price:
          type: string
          description: Price in SAR
        total:
          type: string
          description: Total price in SAR
        qrDataUrl:
          type: string
          description: Base64 data URL of the QR code image
      required:
        - position
        - qrDataUrl
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message