# lib/api-zod/src/generated/types/positionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Asset Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface PositionItem {
  /** Position/Number identifier (e.g., "01 / 1") */
  position: string;
  quantity?: string;
  /** Width in mm */
  width?: string;
  /** Height in mm */
  height?: string;
  /** Area in m²/Pcs. */
  area?: string;
  /** Perimeter in m/Pcs. */
  perimeter?: string;
  /** Price in SAR */
  price?: string;
  /** Total price in SAR */
  total?: string;
  /** Base64 data URL of the QR code image */
  qrDataUrl: string;
}